requirements.txt
akshare>=1.12.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0